Add Vitest CI, co-locate tests, and document contributor workflow
Fix stale calendar/auth tests (frozen clock, alert mock), move specs to
co-located *.test.js, and split calendar coverage across useEvents,
CalendarView, and EventDetailView. Run Vitest on pull requests in CI;
keep Docker build/deploy on pushes to main and dev. Add CONTRIBUTING.md
and update README/CLAUDE for test scripts and conventions.
