+++ package.json
@@ -8,7 +8,8 @@
     "build": "vite-ssg build && node scripts/generate-sitemap.mjs",
     "build:spa": "vite build",
     "preview": "vite preview",
-    "test:unit": "vitest",
+    "test:vitest": "vitest",
+    "test:vitest:ci": "vitest run",
     "test:e2e": "playwright test",
     "test:e2e:ui": "playwright test --ui",
     "test:e2e:headed": "playwright test --headed",
