+++ src/components/LoginPage.test.js
@@ -1,11 +1,11 @@
 import { describe, it, expect } from 'vitest'
 import { mount } from '@vue/test-utils'
-import LoginPage from '../LoginPage.vue'
+import LoginPage from './LoginPage.vue'
 
 describe('LoginPage', () => {
   it('renders properly', () => {
     const wrapper = mount(LoginPage)
-    expect(wrapper.find('h1').text()).toBe('Login')
+    expect(wrapper.find('h1').text()).toBe('Log in with')
   })
 
   it('submits the form with correct data', async () => {
@@ -15,7 +15,6 @@ describe('LoginPage', () => {
     await wrapper.find('#password').setValue('password123')
     await wrapper.find("button[type='submit']").trigger('click')
 
-    // Check if the mock server handled the request correctly
     expect(wrapper.text()).not.toContain('Invalid credentials')
   })
 })
