+++ src/components/jobs/JobListing.test.js
@@ -1,6 +1,8 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest'
 import { mount, flushPromises } from '@vue/test-utils'
-import JobListing from '../jobs/JobListing.vue'
+import { createHead, renderDOMHead } from '@unhead/vue/client'
+import { VueHeadMixin } from '@unhead/vue'
+import JobListing from './JobListing.vue'
 
 const mockJob = {
   id: 'job1',
@@ -12,14 +14,14 @@ const mockJob = {
   how_to_apply: '<p>Email <a href="mailto:hr@acme.com">hr@acme.com</a></p>',
   created: '2025-01-10T12:00:00.000Z',
   approved_at: '2025-01-15T12:00:00.000Z',
-  approved: true,
+  approved: true
 }
 
 function createMockPocketBase(job = mockJob) {
   return {
     collection: vi.fn().mockReturnValue({
-      getOne: vi.fn().mockResolvedValue(job),
-    }),
+      getOne: vi.fn().mockResolvedValue(job)
+    })
   }
 }
 
@@ -29,17 +31,17 @@ function mountJobListing(pb, routeQuery = { id: 'job1' }) {
       config: {
         globalProperties: {
           pocketbase: pb,
-          $route: { query: routeQuery },
-        },
+          $route: { query: routeQuery }
+        }
       },
       stubs: {
         'b-container': { template: '<div><slot /></div>' },
         'b-row': { template: '<div><slot /></div>' },
         'b-col': { template: '<div><slot /></div>' },
         'b-card': { template: '<div class="b-card"><slot /></div>' },
-        'b-badge': { template: '<span class="b-badge"><slot /></span>' },
-      },
-    },
+        'b-badge': { template: '<span class="b-badge"><slot /></span>' }
+      }
+    }
   })
 }
 
@@ -101,7 +103,7 @@ describe('JobListing', () => {
   it('sanitizes HTML in description', async () => {
     const job = {
       ...mockJob,
-      description: '<p>Hello</p><script>alert("xss")</script>',
+      description: '<p>Hello</p><script>alert("xss")</script>'
     }
     const localPb = createMockPocketBase(job)
     const wrapper = mountJobListing(localPb)
@@ -115,7 +117,7 @@ describe('JobListing', () => {
   it('sanitizes HTML in how_to_apply', async () => {
     const job = {
       ...mockJob,
-      how_to_apply: '<p>Apply</p><img src=x onerror=alert(1)>',
+      how_to_apply: '<p>Apply</p><img src=x onerror=alert(1)>'
     }
     const localPb = createMockPocketBase(job)
     const wrapper = mountJobListing(localPb)
@@ -143,5 +145,46 @@ describe('JobListing', () => {
     expect(pb.collection).not.toHaveBeenCalled()
     expect(consoleSpy).toHaveBeenCalledWith('No jobId provided in query parameters.')
     consoleSpy.mockRestore()
+  })
+
+  it('emits valid JobPosting JSON-LD once the job loads', async () => {
+    const head = createHead()
+    const job = {
+      id: 'job1',
+      title: 'Senior Developer',
+      company: 'Acme Corp',
+      salary_min: 100,
+      salary_max: 150,
+      description: '<p>Great <strong>opportunity</strong></p>',
+      how_to_apply: '',
+      created: '2025-01-10T12:00:00.000Z',
+      approved_at: '2025-01-15T12:00:00.000Z'
+    }
+    const localPb = { collection: () => ({ getOne: vi.fn().mockResolvedValue(job) }) }
+
+    mount(JobListing, {
+      global: {
+        plugins: [head],
+        mixins: [VueHeadMixin],
+        config: { globalProperties: { pocketbase: localPb, $route: { query: { id: 'job1' } } } },
+        stubs: {
+          'b-card': { template: '<div><slot /></div>' },
+          'b-badge': { template: '<span><slot /></span>' }
+        }
+      }
+    })
+    await flushPromises()
+    await renderDOMHead(head)
+
+    const tag = document.querySelector('script[type="application/ld+json"]')
+    expect(tag).toBeTruthy()
+    const data = JSON.parse(tag.textContent)
+    expect(data['@type']).toBe('JobPosting')
+    expect(data.title).toBe('Senior Developer')
+    expect(data.hiringOrganization.name).toBe('Acme Corp')
+    expect(data.datePosted).toBe('2025-01-15')
+    expect(data.baseSalary.value.minValue).toBe(100000)
+    expect(data.baseSalary.value.maxValue).toBe(150000)
+    expect(data.jobLocation.address.addressRegion).toBe('IN')
   })
 })
