+++ src/seo.test.js
@@ -1,8 +1,7 @@
-import { describe, it, expect, afterEach, vi } from 'vitest'
+import { describe, it, expect, afterEach } from 'vitest'
 import { mount, flushPromises } from '@vue/test-utils'
 import { createHead, renderDOMHead } from '@unhead/vue/client'
 import { VueHeadMixin } from '@unhead/vue'
-import JobListing from '@/components/jobs/JobListing.vue'
 import {
   SITE_URL,
   SITE_NAME,
@@ -42,7 +41,6 @@ describe('seo helpers', () => {
     expect(tag.key).toBe('ld-x')
     expect(tag.innerHTML).not.toContain('</script>')
     expect(tag.innerHTML).toContain('\\u003c')
-    // Round-trips back to the original object (the escape is JSON-safe).
     expect(JSON.parse(tag.innerHTML).name).toBe('</script><b>')
   })
 
@@ -54,8 +52,6 @@ describe('seo helpers', () => {
   })
 })
 
-// Verifies the actual unhead wiring used in main.js: createHead() + the
-// VueHeadMixin that powers the Options-API `head()` option App.vue relies on.
 describe('unhead head() wiring', () => {
   afterEach(() => {
     document.head.innerHTML = ''
@@ -67,8 +63,6 @@ describe('unhead head() wiring', () => {
     head = createHead()
     return mount(component, { global: { plugins: [head], mixins: [VueHeadMixin] } })
   }
-  // unhead's client renderer is rAF-debounced; force a synchronous flush so the
-  // assertions don't race the DOM write.
   const renderHead = () => renderDOMHead(head)
 
   it('renders a reactive head() option to the document', async () => {
@@ -96,7 +90,6 @@ describe('unhead head() wiring', () => {
       `About · ${SITE_NAME}`
     )
 
-    // Reactivity: changing state updates the rendered tags.
     wrapper.vm.title = 'Jobs'
     wrapper.vm.description = 'Job openings'
     await flushPromises()
@@ -104,46 +97,5 @@ describe('unhead head() wiring', () => {
 
     expect(document.title).toBe(`Jobs · ${SITE_NAME}`)
     expect(document.querySelector('meta[name="description"]')?.content).toBe('Job openings')
-  })
-
-  it('JobListing emits valid JobPosting JSON-LD once the job loads', async () => {
-    head = createHead()
-    const job = {
-      id: 'job1',
-      title: 'Senior Developer',
-      company: 'Acme Corp',
-      salary_min: 100,
-      salary_max: 150,
-      description: '<p>Great <strong>opportunity</strong></p>',
-      how_to_apply: '',
-      created: '2025-01-10T12:00:00.000Z',
-      approved_at: '2025-01-15T12:00:00.000Z'
-    }
-    const pb = { collection: () => ({ getOne: vi.fn().mockResolvedValue(job) }) }
-
-    mount(JobListing, {
-      global: {
-        plugins: [head],
-        mixins: [VueHeadMixin],
-        config: { globalProperties: { pocketbase: pb, $route: { query: { id: 'job1' } } } },
-        stubs: {
-          'b-card': { template: '<div><slot /></div>' },
-          'b-badge': { template: '<span><slot /></span>' }
-        }
-      }
-    })
-    await flushPromises()
-    await renderHead()
-
-    const tag = document.querySelector('script[type="application/ld+json"]')
-    expect(tag).toBeTruthy()
-    const data = JSON.parse(tag.textContent)
-    expect(data['@type']).toBe('JobPosting')
-    expect(data.title).toBe('Senior Developer')
-    expect(data.hiringOrganization.name).toBe('Acme Corp')
-    expect(data.datePosted).toBe('2025-01-15')
-    expect(data.baseSalary.value.minValue).toBe(100000)
-    expect(data.baseSalary.value.maxValue).toBe(150000)
-    expect(data.jobLocation.address.addressRegion).toBe('IN')
   })
 })
